Updating the Pebble Time models
We have asked our hardware team to re-export the models for Pebble Time
making sure that they are manifold, that they are laying down on the x/y
axis (for easier printing out of the box) while still maintaining
accuracy and quality of the file.

This should close #2 and #3. If for some reasons you are unhappy with
them please tell us what software you are using and what problems you
are seeing. Thanks for all the feedback!